go-concourse: structure: build events uses visitor
There's not much benefit to using a visitor here, since there's only one
event type. However, other streams that can produce multiple event types
will benefit from the strong typing.

Note that this is a breaking change to go-concourse. I'm not sure what
our policy is on maintaining stability in the go-concourse API, but it
might make sense to keep the existing signature and deprecate it in
favour of the visitor pattern.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>